Se agrega frmProfecional y se prueba su funcionamiento

+++ dominio/Profesional.cs
@@ -11,7 +11,7 @@ namespace dominio
         public int IdProfesional { get; set; }
         public int Matricula {  get; set; }
         public Especialidad Especialidad { get; set; }
-        public float Honorario { get; set; }
+        public decimal Honorario { get; set; }
         public int Dni { get; set; }
         public bool Activo { get; set; }
